requirements.txt
opencv-contrib-python
opencv-python
dlib
face-recognition
mtcnn
retina-face
customtkinter
CTkListbox
CTkMessagebox
Pillow
pymediainfo
pyinstaller